requirements.txt
Django==4.2.6
openpyxl==3.1.2
pandas==2.1.1
gunicorn
whitenoise
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-filter==23.5
django-widget-tweaks==1.5.0

